+++ lib/cmds/merge_cmds/export.ts
@@ -1,10 +1,8 @@
 import type { Argv } from 'yargs'
 import { handleAsyncError as handle } from '../../utils/async'
-import { success, warning } from '../../utils/log'
-import { confirmation } from '../../utils/actions'
-import { installApp, isAppInstalled } from '../../utils/app-installation'
+import { success } from '../../utils/log'
 import { createManagementClient } from '../../utils/contentful-clients'
-import { type ClientAPI } from 'contentful-management'
+import { checkAndInstallAppInEnvironments } from '../../utils/app-installation'
 
 const MERGE_APP_ID = 'cQeaauOu1yUCYVhQ00atE'
 
@@ -50,101 +48,6 @@ interface ExportMigrationOptions {
   targetEnvironmentId: string
   yes?: boolean
   outputFile?: string
-}
-
-const promptAppInstallationInEnvironment = async (
-  client: ClientAPI,
-  spaceId: string,
-  environmentId: string,
-  appId: string
-) => {
-  warning(
-    `The Merge app is not installed in the environment with id: ${environmentId}`
-  )
-
-  const userConfirmation = await confirmation(
-    `Do you want to install the Merge app in the environment with id: ${environmentId}`
-  )
-
-  if (userConfirmation) {
-    return false
-  }
-
-  await installApp(client, {
-    spaceId,
-    environmentId: environmentId,
-    appId
-  })
-
-  return true
-}
-
-export const checkAndInstallAppInEnvironments = async (
-  client: ClientAPI,
-  spaceId: string,
-  environmentIds: [string, string],
-  appId: string,
-  continueWithoutPrompt: boolean
-) => {
-  const appInstallations = {
-    source: await isAppInstalled(client, {
-      spaceId: spaceId,
-      environmentId: environmentIds[0],
-      appId: MERGE_APP_ID
-    }),
-    target: await isAppInstalled(client, {
-      spaceId: spaceId,
-      environmentId: environmentIds[1],
-      appId: MERGE_APP_ID
-    })
-  }
-
-  if (appInstallations.source && appInstallations.target) {
-    return true
-  }
-
-  // User has passed the --yes flag
-  if (continueWithoutPrompt) {
-    // Install the app in both environments. If it's already installed it will just continue.
-    await installApp(client, {
-      spaceId,
-      environmentId: environmentIds,
-      appId
-    })
-  } else {
-    if (!appInstallations.source && !appInstallations.target) {
-      warning(
-        `The Merge app is not installed in any of the environments. Environment ids: ${environmentIds[0]}, ${environmentIds[1]}`
-      )
-      const userConfirmation = await confirmation(
-        `Do you want to install the Merge app in both environments?`
-      )
-
-      if (!userConfirmation) {
-        return false
-      }
-
-      await installApp(client, {
-        spaceId,
-        environmentId: environmentIds,
-        appId
-      })
-    } else {
-      for (const env of environmentIds) {
-        const prompt = await promptAppInstallationInEnvironment(
-          client,
-          spaceId,
-          env,
-          MERGE_APP_ID
-        )
-
-        if (!prompt) {
-          return false
-        }
-      }
-    }
-  }
-  return true
 }
 
 const exportEnvironmentMigration = async ({
